Minor casts and format changes to shut up my compilers...

+++ emu/emuloop.c
@@ -1480,11 +1480,11 @@ contcase:     /* the main loop */
     ADVANCE_PC(size_xxxXX);
     XSB_Deref(op1); 
     if (isnumber(op1)) {
-      if (int_val(op1) == op2)
+      if (int_val(op1) == (Integer)op2)
 	lpcreg = (byte *)op3;
     }
     else if (isboxedinteger(op1)) {
-       if (oint_val(op1) == op2)
+       if (oint_val(op1) == (Integer)op2)
           lpcreg = (byte *)op3;
     }	  
     else if (isboxedfloat(op1)) {
@@ -1504,11 +1504,11 @@ contcase:     /* the main loop */
     ADVANCE_PC(size_xxxXX);
     XSB_Deref(op1); 
     if (isnumber(op1)) {
-      if (int_val(op1) != op2)
+      if (int_val(op1) != (Integer)op2)
 	lpcreg = (byte *) op3;
     }
     else if (isboxedinteger(op1)) {
-       if (oint_val(op1) != op2)
+       if (oint_val(op1) != (Integer)op2)
           lpcreg = (byte *)op3;
     }	  
     else if (isboxedfloat(op1)) {
